// WebApplicationPlateforme/ClientApp/src/app/shared/Models/Orphelin/psy-orphelin.model.ts
export class PsyOrphelin {

  id: number;
        maladiePsy: string;
        type: string;
        docteur: string;
        periode: string;
        besoinsDoc: string;
        periodeDod: string;


attribut1: number;
attribut7: number;
        attribut2: string;
        attribut3: string;
        attribut4: string;
        attribut5: string;
        attribut6: string;
        dateenreg: string;

idOrph: number;
}